HAAR-830: Use update role admin type in external users api - instead of auth
