+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/service/ExternalUsersApiService.kt
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.Role
+import uk.gov.justice.digital.hmpps.manageusersapi.resource.RoleAdminTypeAmendment
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.RolesPaged
 import java.time.Duration
 
@@ -63,6 +64,20 @@ class ExternalUsersApiService(
       .retrieve()
       .bodyToMono(Role::class.java)
       .block(timeout) ?: throw RoleNotFoundException("get", roleCode, "notfound")
+
+  @Throws(RoleNotFoundException::class)
+  fun updateRoleAdminType(roleCode: String, roleAmendment: RoleAdminTypeAmendment) {
+    log.debug("Updating role for {} with {}", roleCode, roleAmendment)
+    externalUsersWebClient.put()
+      .uri("/api/roles/$roleCode/admintype")
+      .bodyValue(mapOf("adminType" to roleAmendment.adminType.addDpsAdmTypeIfRequiredAsList()))
+      .retrieve()
+      .toBodilessEntity()
+      .block(timeout)
+  }
 }
+
+private fun Set<AdminType>.addDpsAdmTypeIfRequiredAsList() =
+  (if (AdminType.DPS_LSA in this) (this + AdminType.DPS_ADM) else this).map { it.adminTypeCode }
 
 class RoleList : MutableList<Role> by ArrayList()
