+++ src/test/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/integration/wiremock/HmppsAuthMockServer.kt
@@ -109,28 +109,6 @@ class HmppsAuthMockServer : WireMockServer(WIREMOCK_PORT) {
     )
   }
 
-  fun stubPutRoleAdminType(roleCode: String) {
-    stubFor(
-      put("/auth/api/roles/$roleCode/admintype")
-        .willReturn(
-          aResponse()
-            .withStatus(HttpStatus.OK.value())
-            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
-        )
-    )
-  }
-
-  fun stubPutRoleAdminTypeFail(roleCode: String, status: HttpStatus) {
-    stubFor(
-      put("/auth/api/roles/$roleCode/admintype")
-        .willReturn(
-          aResponse()
-            .withHeader("Content-Type", "application/json")
-            .withStatus(status.value())
-        )
-    )
-  }
-
   fun stubForNewToken() {
     stubFor(
       post(urlEqualTo("/auth/api/new-token"))
